# fix sql compatible null handling config work with runtime.properties