--- a/core/src/main/java/org/apache/druid/common/config/NullHandling.java
+++ b/core/src/main/java/org/apache/druid/common/config/NullHandling.java
@@ -50,16 +50,21 @@ public class NullHandling
    * It does not take effect in all unit tests since we don't use Guice Injection.
    */
   @Inject
-  private static NullValueHandlingConfig INSTANCE = new NullValueHandlingConfig(
-      Boolean.valueOf(System.getProperty(NULL_HANDLING_CONFIG_STRING, "true"))
-  );
+  private static NullValueHandlingConfig INSTANCE;
+
+  private static Boolean isReplaceWithDefault = null;
 
   /**
    * whether nulls should be replaced with default value.
    */
   public static boolean replaceWithDefault()
   {
-    return INSTANCE.isUseDefaultValuesForNull();
+    if (isReplaceWithDefault == null) {
+      isReplaceWithDefault = INSTANCE != null
+                             ? INSTANCE.isUseDefaultValuesForNull()
+                             : Boolean.parseBoolean(System.getProperty(NULL_HANDLING_CONFIG_STRING, "true"));
+    }
+    return isReplaceWithDefault;
   }
 
   public static boolean sqlCompatible()